requirements-dev.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
pytest>=7.0.0
pylint>=2.15.0
black>=22.10.0
mypy>=0.991
boto3>=1.26.0
python-dotenv>=0.21.0
pytest-cov>=4.0.0
aws-cdk.aws-apigatewayv2-alpha>=2.0.0
types-requests>=2.28.0
pydantic>=1.10.0
